Fixes errors and checks that all examples are functioning
