+++ dl_workshop/cost_functions.py
@@ -4,20 +4,23 @@ def binary_crossentropy(y_true, y_pred, deriv=False, eps=1e-12):
     """
     Binary Crossentropy
     """
-    if deriv:
-        # Prevent overflow
-        outputs = np.clip(y_pred, eps, 1 - eps)
-        divisor = np.maximum(outputs * (1 - outputs), eps)
-        return (y_pred - y_true) / divisor
-    return -np.sum(np.multiply(y_true, np.log(y_pred)) + np.multiply((1 - y_true), np.log(1 - y_pred))) / y_true.shape[1]
+    # Prevent overflow
+    outputs = np.clip(y_pred, eps, 1 - eps)
+    divisor = np.maximum(outputs * (1 - outputs), eps)
 
-def categorical_crossentropy(y_true, y_pred, deriv=False, epsilon=1e-12):
+    if deriv:
+        return (y_pred - y_true) / divisor
+    return -np.sum(np.multiply(y_true, np.log(outputs)) + np.multiply((1 - y_true), np.log(1 - outputs))) / y_true.shape[1]
+
+def categorical_crossentropy(y_true, y_pred, deriv=False, eps=1e-12):
     """
     Categorical Crossentropy
     """
-    y_pred = np.clip(y_pred, epsilon, 1 - epsilon)
-    
+    # Prevent overflow
+    outputs = np.clip(y_pred, eps, 1 - eps)
+    divisor = np.maximum(outputs * (1 - outputs), eps)
+
     if deriv:
-        return y_pred - y_true
+        return (y_pred - y_true) / divisor
     else:
         return -np.sum(np.multiply(y_true, np.log(y_pred))) / y_true.shape[1]
